style: 패키지 구조 변경하면서 파일 정리
