+++ src/main/java/com/kreamish/kream/common/config/QueryDslConfig.java
@@ -1,14 +1,12 @@
-package com.kreamish.kream.config;
+package com.kreamish.kream.common.config;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import jakarta.persistence.*;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
 @Configuration
-@EnableJpaAuditing
 public class QueryDslConfig {
 
     @PersistenceContext
