+++ src/main/java/com/kreamish/kream/legacy/entity/Collection.java
@@ -1,8 +1,17 @@
-package com.kreamish.kream.entity;
+package com.kreamish.kream.legacy.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.hibernate.validator.constraints.Length;
 
 @Entity
@@ -11,14 +20,16 @@ import org.hibernate.validator.constraints.Length;
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @EqualsAndHashCode
 public class Collection extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long collectionId;
 
-    @Length(min=1,max=50)
+    @Length(min = 1, max = 50)
     @Column(name = "name", nullable = false, unique = true)
     @Setter
     private String name;
+
     public static Collection of(String name) {
         return new Collection(null, name);
     }
