+++ src/main/java/com/kreamish/kream/legacy/entity/Comment.java
@@ -1,26 +1,30 @@
-package com.kreamish.kream.entity;
+package com.kreamish.kream.legacy.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class MemberRole extends BaseEntity {
+public class Comment extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long memberRoleId;
+    private Long commentId;
 
-    @Length(max = 16)
-    @Column(nullable = false)
-    private String name;
+    @ManyToOne
+    @JoinColumn(name = "item_Id")
+    private Item item;
+
+    @ManyToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
 }
