+++ src/main/java/com/kreamish/kream/legacy/entity/ItemCollectionRel.java
@@ -1,4 +1,4 @@
-package com.kreamish.kream.entity;
+package com.kreamish.kream.legacy.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import jakarta.persistence.Entity;
