+++ src/main/java/com/kreamish/kream/legacy/entity/MemberRole.java
@@ -1,4 +1,4 @@
-package com.kreamish.kream.entity;
+package com.kreamish.kream.legacy.entity;
 
 import com.kreamish.kream.common.entity.BaseEntity;
 import jakarta.persistence.Column;
@@ -6,8 +6,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,21 +14,13 @@ import org.hibernate.validator.constraints.Length;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member extends BaseEntity {
+public class MemberRole extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long memberId;
+    private Long memberRoleId;
 
-    @ManyToOne
-    @JoinColumn(name = "member_role_id")
-    private MemberRole memberRole;
-
+    @Length(max = 16)
     @Column(nullable = false)
-    @Length(max = 200)
-    private String email;
-
-    @Column(nullable = false)
-    @Length(max = 256)
-    private String password;
+    private String name;
 }
