+++ src/main/java/com/kreamish/kream/legacy/repository/BrandRepository.java
@@ -1,7 +1,8 @@
-package com.kreamish.kream.repository;
+package com.kreamish.kream.legacy.repository;
 
-import com.kreamish.kream.entity.Brand;
+import com.kreamish.kream.legacy.entity.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BrandRepository extends JpaRepository<Brand, Long>, ItemQueryRepository {
+
 }
