+++ src/main/java/com/kreamish/kream/legacy/repository/CollectionRepository.java
@@ -1,8 +1,8 @@
-package com.kreamish.kream.repository;
+package com.kreamish.kream.legacy.repository;
 
-import com.kreamish.kream.entity.Collection;
-import com.kreamish.kream.entity.Item;
+import com.kreamish.kream.legacy.entity.Collection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CollectionRepository extends JpaRepository<Collection, Long>, ItemQueryRepository {
+
 }
