+++ src/main/java/com/kreamish/kream/legacy/repository/ItemCollectionRelRepository.java
@@ -1,0 +1,9 @@
+package com.kreamish.kream.legacy.repository;
+
+import com.kreamish.kream.legacy.entity.ItemCollectionRel;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemCollectionRelRepository extends JpaRepository<ItemCollectionRel, Long>,
+    ItemQueryRepository {
+
+}
