+++ src/main/java/com/kreamish/kream/legacy/repository/ItemRepository.java
@@ -1,7 +1,8 @@
-package com.kreamish.kream.repository;
+package com.kreamish.kream.legacy.repository;
 
-import com.kreamish.kream.entity.Item;
+import com.kreamish.kream.legacy.entity.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemRepository extends JpaRepository<Item, Long>, ItemQueryRepository {
+
 }
