+++ src/main/java/com/kreamish/kream/legacy/repository/ItemSizesRepository.java
@@ -1,6 +1,6 @@
-package com.kreamish.kream.repository;
+package com.kreamish.kream.legacy.repository;
 
-import com.kreamish.kream.entity.ItemSizes;
+import com.kreamish.kream.legacy.entity.ItemSizes;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ItemSizesRepository extends JpaRepository<ItemSizes, Long>, ItemQueryRepository {
