--- src/main/java/com/kreamish/kream/repository/ItemCollectionRelRepository.java
@@ -1,7 +1,0 @@
-package com.kreamish.kream.repository;
-
-import com.kreamish.kream.entity.ItemCollectionRel;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ItemCollectionRelRepository extends JpaRepository<ItemCollectionRel, Long>, ItemQueryRepository {
-}
